% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compare_teams}
\alias{compare_teams}
\title{Compare Teams}
\usage{
compare_teams(team1_name, team2_name)
}
\arguments{
\item{team1_name}{The name of the first soccer team for comparison.}

\item{team2_name}{The name of the second soccer team for comparison.}
}
\value{
A spider plot comparing the performance metrics of the two teams.
}
\description{
Compares the performance metrics of two soccer teams in the English Premier League during the 2023 season.
}
\examples{
compare_teams('Arsenal', 'Manchester United')

}
