% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{predict_final_standings}
\alias{predict_final_standings}
\title{Function 3 Title}
\usage{
predict_final_standings()
}
\value{
What the function returns.
}
\description{
Description of what the function does.
}
\examples{
predict_final_standings()



}
