% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{predict_win}
\alias{predict_win}
\title{Predict Win}
\usage{
predict_win()
}
\value{
A data frame containing match details and predicted outcomes.
}
\description{
Predicts the outcomes of the remaining matches in the Premier League 2023-24 season.
}
\details{
This function retrieves football match data from API - Football, by training a linear regression model on historical data, and predicts match outcomes for upcoming fixtures.
}
\examples{
match_data <- predict_win()
head(match_data)


}
