% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{top_20}
\alias{top_20}
\title{Top 20}
\usage{
top_20(df)
}
\value{
The stacked bar plot visualizing the top 20 EPL players.
}
\description{
Visualize thee top 20 EPL players of the 23-24 season based on goals and assists. (This function is not intended for user access.)
}
\details{
This function takes a data frame containing statistics of EPL players and visualizes the top
20 players based on their combined goals and assists. It creates a bar chart using the Plotly
library, with players' names reordered based on their combined goals and assists.
}
\examples{
# This function is not meant to be called directly by users.

}
\keyword{internal}
