% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{track_performance}
\alias{track_performance}
\title{Function 2 Title}
\usage{
track_performance(team_name)
}
\value{
What the function returns.
}
\description{
Description of what the function does.
}
\examples{
track_performance('Arsenal')


}
