% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compare_teams}
\alias{compare_teams}
\title{Function 1 Title}
\usage{
compare_teams(team1_name, team2_name)
}
\value{
What the function returns.
}
\description{
Description of what the function does.
}
\examples{
compare_teams('Arsenal', 'Manchester United')


}
