% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{predict_final_standings}
\alias{predict_final_standings}
\title{Predict Final Standings}
\usage{
predict_final_standings()
}
\value{
A data frame containing the final league standings.
}
\description{
Predicts the final league standings of the Premier League 2023-24 season.
}
\details{
This function combines match prediction results with current league standings to generate the final league table. It retrieves match data and standings from specified APIs, performs calculations, and produces a visual representation of the league table.
}
\examples{
final_standings <- predict_final_standings()
head(final_standings)



}
