% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashkick.R
\name{top_20}
\alias{top_20}
\title{Top 20 Function (Internal)}
\usage{
top_20()
}
\value{
What the function returns.
}
\description{
Description of what the function does. (This function is not intended for user access.)
}
\examples{
# This function is not meant to be called directly by users.

}
\keyword{internal}
