% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{track_performance}
\alias{track_performance}
\title{Track Performance}
\usage{
track_performance(team_name)
}
\arguments{
\item{team_name}{The name of the soccer team for which performance data is to be tracked.}
}
\value{
A plot displaying the team's performance over the season.
}
\description{
Retrieves and visualizes the performance of a specified soccer team in the English Premier League during the 2023 season.
}
\examples{
track_performance('Arsenal')

}
